+++ Flow/Changes.md
@@ -26,12 +26,15 @@
   - **Wait** now properly breaks (stops) when it encounters an **InterruptedException**
   - `Wait.apply(...)` now returns a boolean that indicates whether the wait was forcibly interrupted or not
   - Modified the following classes to support interruptions, also:
-    - **DelayedView**
+    - **DelayedProcess**
     - **LoopingProcess**
     - **TimedTasks**
     - **ExpiringCache**
     - **ExpiringLazy**
+    - **DelayedView**
     - **KeptOpenWriter**
+  - **Future**`.raceWith(Future)` may now yield a failing future in case the process is interrupted with an 
+    **InterruptedException**
 - Rewrote **Value**`.castTo(DataType, DataType)` so that it will cast to the closer data type
 - **Volatile**`.value` is no longer synchronized. For synchronized access, use `.synchronizedValue`
 - **VolatileFlag**.`set()` and `.reset()` now return booleans that indicate whether the flag state was actually modified